Moved dir_entry to directory
